+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Tucan.Finance.candlestick(:ohlc, "date", "open", "high", "low", "close",
 )
 ```
 
+- Add `Tucan.Axes.set_color/3` helper.
+
 ## [v0.3.1](https://github.com/pnezis/tucan/tree/v0.3.1) (2024-01-20)
 
 ### Added plot options